TINY-8588: Fixups
